fixed import errors due to refactoring of craco
