+++ cryptoprotocol/src/main/java/de/upb/crypto/incentive/cryptoprotocol/model/keys/user/UserKeyPair.java
@@ -1,6 +1,5 @@
 package de.upb.crypto.incentive.cryptoprotocol.model.keys.user;
 
-
 public class UserKeyPair {
     private UserPublicKey userPublicKey;
     private UserSecretKey userSecretKey;
