+++ cryptoprotocol/src/main/java/de/upb/crypto/incentive/cryptoprotocol/model/keys/user/UserPublicKey.java
@@ -1,6 +1,6 @@
 package de.upb.crypto.incentive.cryptoprotocol.model.keys.user;
 
-import de.upb.crypto.math.interfaces.structures.GroupElement;
+import de.upb.crypto.math.structures.groups.GroupElement;
 
 public class UserPublicKey {
     private GroupElement upk;
