+++ cryptoprotocol/src/main/java/de/upb/crypto/incentive/cryptoprotocol/model/messages/EarnRequestMessage.java
@@ -1,7 +1,7 @@
 package de.upb.crypto.incentive.cryptoprotocol.model.messages;
 
 import de.upb.crypto.craco.sig.sps.eq.SPSEQSignature;
-import de.upb.crypto.math.interfaces.structures.GroupElement;
+import de.upb.crypto.math.structures.groups.GroupElement;
 
 public class EarnRequestMessage {
     private GroupElement blindedCommitment;
