+++ cryptoprotocol/src/main/java/de/upb/crypto/incentive/cryptoprotocol/protocols/provider/EarnRequestHandler.java
@@ -1,12 +1,14 @@
 package de.upb.crypto.incentive.cryptoprotocol.protocols.provider;
 
 import de.upb.crypto.incentive.cryptoprotocol.interfaces.provider.CreditInterface;
+import de.upb.crypto.math.serialization.Representable;
+import de.upb.crypto.math.serialization.Representation;
 
 /**
  * A handler class implementing the provider side of the Credit <-> Earn protocol (i.e. the Credit algorithm).
  * Handles Earn requests made by users.
  */
-public class EarnRequestHandler implements CreditInterface {
+public class EarnRequestHandler implements CreditInterface, Representable {
     public String computeSerializedResponse(String serializedEarnRequest, long earnAmount) {
         // deserialize earn request
 
@@ -19,4 +21,7 @@ public class EarnRequestHandler implements CreditInterface {
         return null;
     }
 
+    public Representation getRepresentation() {
+        return null;
+    }
 }
