+++ cryptoprotocol/src/main/java/de/upb/crypto/incentive/cryptoprotocol/protocols/user/EarnRequest.java
@@ -7,13 +7,13 @@ import de.upb.crypto.incentive.cryptoprotocol.model.PublicParameters;
 import de.upb.crypto.incentive.cryptoprotocol.model.Token;
 import de.upb.crypto.incentive.cryptoprotocol.model.keys.provider.ProviderPublicKey;
 import de.upb.crypto.incentive.cryptoprotocol.model.keys.user.UserSecretKey;
-import de.upb.crypto.math.interfaces.structures.GroupElement;
 import de.upb.crypto.math.serialization.Representable;
 import de.upb.crypto.math.serialization.Representation;
-import de.upb.crypto.math.serialization.annotations.v2.ReprUtil;
-import de.upb.crypto.math.serialization.annotations.v2.Represented;
-import de.upb.crypto.math.structures.zn.Zn;
-import de.upb.crypto.math.structures.zn.Zn.ZnElement;
+import de.upb.crypto.math.serialization.annotations.ReprUtil;
+import de.upb.crypto.math.serialization.annotations.Represented;
+import de.upb.crypto.math.structures.groups.GroupElement;
+import de.upb.crypto.math.structures.rings.zn.Zn;
+import de.upb.crypto.math.structures.rings.zn.Zn.ZnElement;
 
 /**
  * Class representing an earn request created by a user of the incentive system.
